chore: extract DataRefreshService and TabRegistry from CamelMonitor
Step 4: Extract DataRefreshService (~666 lines) — moves all background
data polling, vanishing lifecycle, auto-select, and infra/trace/span/error
data loading out of CamelMonitor. Uses RefreshContext callback interface
(6 methods) to decouple from monitor state.

Step 5: Extract TabRegistry (~305 lines) — moves tab instance lifecycle,
switching logic, and tab index constants. Uses TabCallbacks interface
(7 methods) for event dispatch back to the monitor. Updates McpFacade
to accept TabRegistry instead of individual tab references.

CamelMonitor reduced from ~2,167 to ~1,568 lines. All 260 tests pass.
